SuperSampling AntiAliasing added.
Serveral random number generator added. 
SuperSampling has two modifiable parameters. <dai_random_05> which decide the range of the graph to be supersampling.  <supersamplingnumber> determine the samples.

Trying to find the source of colorful noise. Found that when all colors are of \lamba (1,1,1), the noise disapear.
Warning: the bounchnumber and supersampling number require huge calculus..